Clean duplicate BOLD records that were mined from the NCBI multiple times

This is a convenient function that allows cleaning duplicate BOLD records that were mined from the NCBI but multiple times and thus have a different processID. In those cases, the accession numbers those records reports will be the same, thus the original NCBI record is searched using different utils functions listed in utils_helpers.R to search those records and check the length of the original CDS/rRNA.

Create, from the records object, a new table that includes two new columns one corresponding to the combination of processID and gene name that is also supplied to this function as accn, and another with the corresponding accession number. The latter is created using the coalesce function of dplyr which allows to merge two vectors, in this case columns of a table, replacing the NA items with the corresponding value, if present, in the other column. Important to remember that coalesce works giving preference to the first vector supplied, thus if both columns present the value the one coming from the column NCBI_ID will be used instead of the other. This allows to recover all those accession numbers that are not present in the correct field of the record mined from Genbank but instead in the sampleID field. The lengthGene field is quickly calculated using the sequences supplied to the function. Moreover, some records on BOLD have no sequence associated. Those will be removed beforehand.

Usage

cleanMinedFromGenBankRecords(records, sequences, accn)

Arguments

records

data.frame the output of bold fetcher format

sequences

DNAStringSet the sequences corresponding to the records table, but in DNAStringSet format

names

character a vector of processID and corresponding marker codes as selected prevously by the user

Value

character the same as parameter name, but including the correct mined record