Example records data frame

An example data frame, including meta-data for a fake sequence used as example dataset for the function loadBarcodeOre or for the vignettes.

Usage

example_record

Format

## `example_record` A data frame with 1 row and 25 columns:

sourceID

Record code

source

Acronym/name of the database/department from which the record/sequence originates

phylum, class, order, family, genus, species

Taxonomic classification of the record

lengthSource

Length of the sequence, in base pairs

sampleID

Sample code, corresponding to the specimen from which the sequence originates

identified_by, collected_by, taxNotes, collection_date, note

Character class additional informations on the sample's origin

db_xref

Generally unused for user's records, include the taxid of the NCBI record (taxid:"00001") separated by a pipe for the corresponding BOLD processID

NCBI_ID

Accession number of the record on the NCBI database

institutionStoring

Acronym/name of the institution storing the sample

altitude, depth

Elevation data

country

Name of the locality from which the sample originates

lat, lon

Geographical coordinates, in decimal degrees

directionPrimers

Gives the type of primers that are shown in the foeld "PCR_primers": "F|R" indicates that the primers' names in the "PCR_primers" separated by a pipe correspond to the forward and revers primers in this order

PCR_primers

Names of the primers used to amplify the sequence of the record

note

Additional notes

...

Source

generated manually, consult data_raw directory for details